requirements.txt
web3==6.11.1
eth-account==0.9.0
eth-abi==4.2.1
parsimonious==0.9.0
pandas>=1.5.0
numpy>=1.21.0
gymnasium>=0.29.0
python-dotenv==1.0.0
uniswap-python==0.7.1
eth-typing==3.5.1
eth-utils==2.1.0
pytest==7.4.3
seaborn>=0.12.0
matplotlib>=3.7.0
scikit-learn==1.3.0
torch>=2.0.0 